chore: move github call to test_url_checks (WIP)
