+++ .gitignore
@@ -9,3 +9,4 @@ __pycache__
 results/
 venv/
 .DS_store
+~tmp
