+++ requirements/tests.txt
@@ -3,5 +3,6 @@ pytest-variables==1.6.1
 pytest-xdist==1.16.0
 requests==2.14.2
 flake8==3.3.0
+ipgetter
 six
 tox
